+++ src/main/kotlin/com/github/gib/GIdeaBrowserBundle.kt
@@ -5,10 +5,10 @@ import org.jetbrains.annotations.NonNls
 import org.jetbrains.annotations.PropertyKey
 
 @NonNls
-private const val BUNDLE = "messages.MyBundle"
+private const val BUNDLE = "messages.GIdeaBrowserBundle"
 
 @Suppress("unused")
-object MyBundle : AbstractBundle(BUNDLE) {
+object GIdeaBrowserBundle : AbstractBundle(BUNDLE) {
 
     @Suppress("SpreadOperator")
     @JvmStatic