+++ src/main/kotlin/com/github/gib/GivToolWindowFactory.kt
@@ -5,9 +5,6 @@ import com.intellij.openapi.project.DumbAware
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.wm.ToolWindow
 import com.intellij.openapi.wm.ToolWindowFactory
-import com.intellij.openapi.wm.ex.ToolWindowEx
-import com.intellij.openapi.wm.impl.IdeFrameImpl
-import kotlin.math.min
 
 class GivToolWindowFactory : ToolWindowFactory, DumbAware {
 
@@ -22,6 +19,11 @@ class GivToolWindowFactory : ToolWindowFactory, DumbAware {
         val rootComponent = GivMainPanel(GivServiceSettings.instance().getLastSaveHomePage())
         val content = contentManager.factory.createContent(rootComponent, null, false)
         contentManager.addContent(content)
+    }
+
+    companion object{
+        @Suppress("unused")
+        const val GIDEA_BROWSER_TOOL_WINDOW_ID = "GIdeaBrowser"
     }
 
 }