+++ src/main/kotlin/com/github/gib/services/GivServiceSettings.kt
@@ -1,6 +1,5 @@
 package com.github.gib.services
 
-import com.github.gib.SettingsChangedAction
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.components.PersistentStateComponent
 import com.intellij.openapi.components.State
@@ -46,6 +45,7 @@ class GivServiceSettings : PersistentStateComponent<GivServiceSettings.State> {
         }
     }
 
+    @Suppress("unused")
     fun addFavorite(webToFavorite: String) = synchronized(lock) {
         myState.favorites.add(Pair.create(webToFavorite,
             JBImageIcon(ImageIO.read(URL("https://www.google.com/s2/favicons?domain=$webToFavorite")))))